+++ _targets.R
@@ -7,18 +7,20 @@ library("targets")
 if (packageVersion("bittermelon") < "0.1.0-39")
     tar_throw_validate("{bittermelon} too old.  Please upgrade.")
 
+source("R/miscellaneous_symbols.R")
 source("R/box_drawing.R")
 source("R/ttf.R")
 
 tar_option_set(packages = c("bittermelon", "glue", "grid", "hexfont"))
 list(
-    tar_target(version, "0.1.0-4"),
+    tar_target(version, "0.1.0-5"),
     tar_target(font_name, "Game Bit Mono"),
     tar_target(copyright, "Copyright (C) 1998-2021 Trevor L Davis, Roman Czyborra, Paul Hardy, et al. License: SIL Open Font License version 1.1 and GPLv2+: GNU GPL version 2 or later <http://gnu.org/licenses/gpl.html> with the GNU Font Embedding Exception."),
     tar_target(unifont, hexfont::unifont()),
     tar_target(basic_latin, unifont[block2ucp("Basic Latin")]),
     tar_target(latin1_supplement, unifont[block2ucp("Latin-1 Supplement")]),
     tar_target(box_drawing, create_box_drawing()),
+    tar_target(miscellaneous_symbols, create_miscellaneous_symbols(unifont)),
     tar_target(
         block_elements,
         {
@@ -37,14 +39,19 @@ list(
     tar_target(chess_symbols, unifont[block2ucp("Chess Symbols")]),
     tar_target(
         font,
-        c(basic_latin, # U+0020
-          latin1_supplement, # U+00A0
-          box_drawing, # U+2500
-          block_elements, # U+2580
-          cjk_symbols_and_punctuation, # U+3000
-          halfwidth_and_fullwidth_forms, # U+FF01
-          chess_symbols # U+1FA00
-        )
+        {
+            font <- c(basic_latin, # U+0020
+              latin1_supplement, # U+00A0
+              box_drawing, # U+2500
+              block_elements, # U+2580
+              miscellaneous_symbols, # U+2600
+              cjk_symbols_and_punctuation, # U+3000
+              halfwidth_and_fullwidth_forms, # U+FF01
+              chess_symbols # U+1FA00
+            )
+            font <- font[ucp_sort(names(font))]
+            font
+        }
     ),
     tar_target(
         hex_file,